Fix error with AVX CPU feature
Mongo from v5 requires that the CPU has support for AVX.
This prevents the containers from starting on machines without AVX
support.

I have downgraded the version of Mongo to 4.4.6 - a version not
requiring this support.

I also needed to remove command options passed to the mongod client not
supported in this version.